move models file to directory
