+++ backend/app/alembic/env.py
@@ -18,7 +18,7 @@ fileConfig(config.config_file_name)
 # target_metadata = mymodel.Base.metadata
 # target_metadata = None
 
-from app.models import SQLModel  # noqa
+from app.models.models import SQLModel  # noqa
 
 target_metadata = SQLModel.metadata
