+++ backend/app/api/routes/users.py
@@ -12,7 +12,7 @@ from app.api.deps import (
 )
 from app.core.config import settings
 from app.core.security import get_password_hash, verify_password
-from app.models import (
+from app.models.models import (
     Item,
     Message,
     UpdatePassword,
