+++ backend/app/tests/utils/user.py
@@ -3,7 +3,7 @@ from sqlmodel import Session
 
 from app import crud
 from app.core.config import settings
-from app.models import User, UserCreate, UserUpdate
+from app.models.models import User, UserCreate, UserUpdate
 from app.tests.utils.utils import random_email, random_lower_string
